# rlp: use atomic.Value for type cache