@fjl 2021-05-19
I have now added a benchmark to show the performance difference. The benchmark function encodes an `[]interface{}` value on all cores simultaneously, which is the worst case for the old RWLock-based design.

Here's the difference:

    name                         old time/op  new time/op  delta
    EncodeConcurrentInterface-8  4.92µs ± 0%  1.07µs ± 2%  -78.33%  (p=0.000 n=10+9)

The RWLock is clearly visible in the profile before the change:

<img width="629" alt="Screenshot 2021-05-19 at 11 57 57" src="https://user-images.githubusercontent.com/6915/118794026-76594100-b899-11eb-8033-936fab35cb71.png">